class nodes implemented

+++ src/radix-trie/file-info.hpp
@@ -1,5 +1,5 @@
 #pragma once
-
+#include "../global.hpp"
 
 
 namespace pinguqueen {
@@ -7,7 +7,7 @@ namespace pinguqueen {
     struct FileInfo {
 
         std::string file_name;
-        uint64_t file_size_bytes = 0;
+        u32 file_size_bytes = 0;
     //  bool is_directory = false;
 
     };
